+++ femboy-meter/commands/fembucks.js
@@ -1,0 +1,4 @@
+export default function command({bot, db, parsed, message}) {
+    if(!db.data.hasOwnProperty(message.author)) return;
+    bot.post(`You have ${db.data[message.author].fembucks} fembucks!`, message.origin)
+}
