+++ femboy-meter/commands/help.js
@@ -1,0 +1,4 @@
+export default function command({bot, db, parsed, message}) {
+    // if(!db.data.hasOwnProperty(message.author)) return;
+    bot.post('https://twink.pages.dev/femboy\n-# remember to avoid using bots in home! create a gc instead :3', message.origin)
+}
