+++ femboy-meter/commands/stocks.js
@@ -1,0 +1,4 @@
+export default function command({bot, db, parsed, message}) {
+    // if(!db.data.hasOwnProperty(message.author)) return;
+    bot.post(`sell price: ${trade.sell}\nreturn: ${trade.get}`, message.origin)
+}
